+++ src/test/java/jsonvalues/api/gen/TestGenerators.java
@@ -19,9 +19,7 @@ import java.time.ZonedDateTime;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
-import java.util.Set;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 import static jsonvalues.spec.JsSpecs.*;